requirement.txt
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
google-auth==2.37.0
google-genai==0.3.0
grpcio==1.70.0
idna==3.10
mss==10.0.0
numpy==2.2.1
opencv-python==4.10.0.84
pillow==11.0.0
protobuf==3.20.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
python-osc==1.9.3
requests==2.32.3
rsa==4.9
soundfile==0.13.1
SpeechRecognition==3.14.2
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.1
